AK: Make sign_extend() actually work for non-64-bit types
sign_extend() didn't do the right thing for types narrower than 64 bits:
When calling e.g. sign_extend(0x10u, 5), T was an int, so
`static_cast<i64>(value << (sizeof(T) * 8 - bits))` would shift 0x10u by
32 - 5 == 27 to the left, then convert that to an i64 (which has plenty
of room for a 32-bit unsigned), and shift that back.

Now we cast to signed int before the shift, and then shifting back does
sign extension as intended.

This fixes regressions introduced in #25805 in FlacLoader.cpp and
JBIG2Loader.cpp.

Also add some tests. All of these tests that expect negative results
but the last three failed with the old implementation.

+++ AK/IntegralMath.h
@@ -90,7 +90,7 @@ constexpr MakeSigned<T> sign_extend(T value, u8 bits)
 {
     // C++ considers the shift by sizeof(T) * 8 UB, and it doesn’t make logical sense to sign-extend 0 bits anyways.
     VERIFY(bits > 0);
-    return static_cast<MakeSigned<T>>((static_cast<i64>(value << (sizeof(T) * 8 - bits))) >> (sizeof(T) * 8 - bits));
+    return static_cast<MakeSigned<T>>(value << (sizeof(T) * 8 - bits)) >> (sizeof(T) * 8 - bits);
 }
 
 }
